+++ tourcount/src/main/java/com/wmstein/tourcount/database/IndividualsDataSource.kt
@@ -14,6 +14,7 @@ import com.wmstein.tourcount.MyDebug
  * Created by wmstein for TourCount on 2016-04-20,
  * last edited in Java on 2022-03-24,
  * converted to Kotlin on 2023-07-05
+ * last edited on 2023-12-15,
  */
 class IndividualsDataSource(context: Context?) {
     // Database fields
@@ -169,7 +170,7 @@ class IndividualsDataSource(context: Context?) {
         // check for entries in individuals table, which are not there when bulk counts are entered
         return if (!cursor.isAfterLast) {
             individuals = cursorToIndividuals(cursor)
-            if (MyDebug.LOG) Log.e("IndividDataSource: ", "i_Id = " + individuals.id)
+            if (MyDebug.LOG) Log.e("IndividDataSource: ", "172, i_Id = " + individuals.id)
             cursor.close()
             individuals.id
         } else {