+++ tourcount/src/main/java/com/wmstein/tourcount/database/Section.kt
@@ -4,7 +4,8 @@ package com.wmstein.tourcount.database
  * Definitions for table Section
  * Created by wmstein on 2016-02-18,
  * last edited in Java on 2022-03-23,
- * converted to Kotlin on 2023-07-05
+ * converted to Kotlin on 2023-07-05,
+ * last edited on 2023-12-15.
  */
 class Section {
     @JvmField
@@ -20,7 +21,7 @@ class Section {
     @JvmField
     var place: String? = null
     @JvmField
-    var temp = 0
+    var tmp = 0
     @JvmField
     var wind = 0
     @JvmField