+++ tourcount/src/main/java/com/wmstein/tourcount/database/Temp.kt
@@ -2,6 +2,7 @@ package com.wmstein.tourcount.database
 
 /***********************************
  * Definitions for table Temp
+ * (unfortunately 'Temp' is misinterpreted as error but works as expected)
  * Created by wmstein on 2016-05-15,
  * last edited in Java on 2022-03-23,
  * converted to Kotlin on 2023-07-05
@@ -10,9 +11,7 @@ class Temp {
     @JvmField
     var id = 0
     @JvmField
-    var temp_loc // last locality found or entered
-            : String? = null
+    var temp_loc: String? = null  // last locality found or entered
     @JvmField
-    var temp_cnt // for future use
-            = 0
+    var temp_cnt = 0  // for future use
 }